Remove GitHub Maven workflow file
The Maven workflow file in the GitHub folder, which enabled continuous integration with Java and Maven, has been removed. This change implies that the automatic building of a Java project with Maven on pushes and pull requests to the main branch will no longer be possible. Future builds, dependency updates and checks will need to be handled differently.