Integrate Hibernate-Spring persistence
This commit introduces a Spring integrated Hibernate persistence layer for the application. A new PersistenceConfig Java class is added which uses annotated configurations for Hibernate and DataSource properties. Alongside, POM is updated with necessary dependencies, and app's log file is refactored into an application properties file. README is updated with a reference link for FORTH.