Modify WordService and unburden Bootstrap.groovy.
We modified the WordService to add "complex" words into a Dictionary. The offloading class is called "CoreDefinitions.groovy" and it has methods for creating primitive and complex words. It only needs a dictionary name and a wordService injected to do this. In addition it has a void createCoreDictionary() method to bootstrap a dictionary that does not yet exist.