Update WordService to support complex words, refactor Bootstrap
Implemented support for complex words in WordService through a new class, CoreDefinitions.groovy, which allows creation of primitive and complex words. This simplifies the Bootstrap.groovy by moving some of its responsibilities to CoreDefinitions.groovy. The latter includes a method to bootstrap a non-existent dictionary. This is part of an effort to improve code organization and maintainability.