Update return type and improve method implementation
Changed return type of findFirstByNameOrderByCreateDateTimeDesc method in WordRepository to Optional<Word>. Refactored the findByName method in WordService to handle the optional value. Also, uncommented and adjusted code related to creating complex words in CoreDefinitions and made changes to literal handling in ForthCompiler and ForthInterpreter. Added new runtime classes for Colon and Literal.