Enhance compiler to parse integer literals
This update further expands the compiler's functionality, allowing it to parse integer literal tokens. This is accomplished through the addition of the `compileIntegerLiteral(token)` method in affected classes. Furthermore, the visibility of the `canParseToInt(String str)` and `defineNewWord()` methods is increased for usability. Various changes are also made to the 'PlusLoop' class, utilizing the data stack and the return stack.