Rename classes and update functionality for memory operations
The commit introduces the BlockService to handle operations on memory blocks, replacing the operations directly done on the memory in BlockWord and Store classes. In addition, it activates Commented code in Dictionary class to maintain versioning and dates. Minor renaming of some classes and addition to gitignore file are also part of this commit. Finally, the Forth compiler now uses BlockService, and a new feature for database backup is introduced in Save.groovy.