Implement 'create' compiler directive and primitive word
The new 'create' directive and primitive word have been added to allow compiler and runtime to handle 'create' commands. This has led to modifications in CoreDefinitions, AbstractCompilerDirective, as well as creation of CreateC and Create classes. The 'execute' implementation in 'create' directive creates a new word and saves it into the wordService, while the 'execute' method in the primitive word class manages creation of new words, which affects the REPL behavior.