Change return type of execute method in compiler classes
The return type of the execute method in several classes such as DotQuoteC, CompileC and VariableC has been changed from Boolean to Object. This change provides greater flexibility in terms of the output of the method, and aligns these methods with their usage in the application, where a variety of object types might be returned.