Implement State enumeration and modify exception handling
A new file, State.groovy, has been created to enumerate the states 'INTERPRET' and 'COMPILE'. This is now being used in ForthCompiler and ForthRepl. Additionally, exception handling in WordService and ForthInterpreter has been modified. Instead of returning null or a comment about exception throwing, now `ForthException` is properly thrown when a word is not found.